[Feat] adding auth everywhere its needed and deleted not use resolver

+++ back/nest-back/apps/course/src/course/course.controller.ts
@@ -2,17 +2,23 @@ import { Controller } from '@nestjs/common';
 import { MessagePattern, Payload, RmqContext, Ctx } from '@nestjs/microservices';
 import { CourseService } from './course.service';
 import { Course } from '../../../../libs/database/src/entities/course.entity';
+import { NotFoundException, UnauthorizedException } from '@nestjs/common';
 
 @Controller()
 export class CourseController {
     constructor(private readonly courseService: CourseService) {}
 
     @MessagePattern('course.create')
-    async createCourse(@Payload() payload: Partial<Course>, @Ctx() context: RmqContext) {
+    async createCourse(
+        @Payload() payload: { createCourseInput: Partial<Course>; userId: string },
+        @Ctx() context: RmqContext
+    ) {
         const channel = context.getChannelRef();
         const originalMsg = context.getMessage();
+        const { createCourseInput, userId } = payload;
+        const data = { ...createCourseInput, userId };
         try {
-            const result = await this.courseService.create(payload);
+            const result = await this.courseService.create(data);
             channel.ack(originalMsg);
             return result;
         } catch (error) {
@@ -60,16 +66,28 @@ export class CourseController {
         }
     }
 
+
     @MessagePattern('course.delete')
-    async deleteCourse(@Payload() payload: { id: string }, @Ctx() context: RmqContext) {
+    async deleteCourse(@Payload() payload: { id: string, userId: string }, @Ctx() context: RmqContext) {
+        console.log('Deleting course with payload:', payload);
         const channel = context.getChannelRef();
         const originalMsg = context.getMessage();
         try {
+            const course = await this.courseService.findById(payload.id);
+            console.log('Course to delete:', course);
+            if (!course) {
+                channel.ack(originalMsg);
+                throw new NotFoundException('Course not found');
+            }
+            if (course.userId !== payload.userId) {
+                channel.ack(originalMsg);
+                throw new UnauthorizedException('Unauthorized: You are not the creator of this course');
+            }
             const result = await this.courseService.delete(payload.id);
             channel.ack(originalMsg);
             return result;
         } catch (error) {
-            throw error;
+            return error;
         }
     }
 }
