+++ back/nest-back/apps/schedule/src/schedule.controller.ts
@@ -1,4 +1,4 @@
-import { Controller } from '@nestjs/common';
+import { Controller, UnauthorizedException } from '@nestjs/common';
 import { MessagePattern, Payload, RmqContext, Ctx } from '@nestjs/microservices';
 import { ScheduleService } from './schedule.service';
 import { Schedule } from '../../../libs/database/src/entities/schedule.entity';
@@ -8,11 +8,13 @@ export class ScheduleController {
   constructor(private readonly scheduleService: ScheduleService) {}
 
   @MessagePattern('schedule.create')
-  async createSchedule(@Payload() payload: Partial<Schedule>, @Ctx() context: RmqContext) {
+  async createSchedule(@Payload() payload: { createScheduleInput: Partial<Schedule>, userId: string}, @Ctx() context: RmqContext) {
     const channel = context.getChannelRef();
     const originalMsg = context.getMessage();
+    const { createScheduleInput, userId } = payload;
+    const data = { ...createScheduleInput, userId };
     try {
-        const result = await this.scheduleService.createSchedule(payload);
+        const result = await this.scheduleService.createSchedule(data);
         channel.ack(originalMsg);
         return result;
     }
@@ -24,11 +26,15 @@ export class ScheduleController {
   }
 
   @MessagePattern('schedule.findById')
-  async getSchedule(@Payload() payload: { id: string }, @Ctx() context: RmqContext) {
+  async getSchedule(@Payload() payload: { id: string, userId: string }, @Ctx() context: RmqContext) {
     const channel = context.getChannelRef();
     const originalMsg = context.getMessage();
     try {
         const result = await this.scheduleService.getSchedule(payload.id);
+        if(result?.userId !== payload.userId) {
+            channel.ack(originalMsg);
+            throw new UnauthorizedException('You are not authorized to access this schedule');
+        }
         channel.ack(originalMsg);
         return result;
     }
@@ -66,10 +72,15 @@ export class ScheduleController {
   }
 
   @MessagePattern('schedule.delete')
-  async deleteSchedule(@Payload() payload: { id: string }, @Ctx() context: RmqContext) {
+  async deleteSchedule(@Payload() payload: { id: string, userId: string }, @Ctx() context: RmqContext) {
     const channel = context.getChannelRef();
     const originalMsg = context.getMessage();
     try {
+        const schedule = await this.scheduleService.getSchedule(payload.id);
+        if (schedule?.userId !== payload.userId) {
+            channel.ack(originalMsg);
+            throw new UnauthorizedException('Schedule not found');
+        }
         const result = await this.scheduleService.deleteSchedule(payload.id);
         channel.ack(originalMsg);
         return result;
